tutorialdjango/requirements.txt
Django==3.2.9
django-environ==0.8.1
gunicorn==20.1.0
psycopg2==2.9.2
whitenoise==5.3.0
